sidecar/requirements.txt
fastapi
uvicorn[standard]
requests
argostranslate
langdetect
easyocr
Pillow
opencv-python
numpy
